Optimize options calculation
